+++ Wurst Client/src/org/darkstorm/minecraft/gui/theme/simple/SimpleSliderUI.java
@@ -1,19 +1,15 @@
 package org.darkstorm.minecraft.gui.theme.simple;
 
-import static org.lwjgl.opengl.GL11.*;
-
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.Point;
-import java.awt.Rectangle;
-
 import net.minecraft.client.gui.FontRenderer;
-
 import org.darkstorm.minecraft.gui.component.Container;
 import org.darkstorm.minecraft.gui.component.Slider;
 import org.darkstorm.minecraft.gui.theme.AbstractComponentUI;
 import org.darkstorm.minecraft.gui.util.RenderUtil;
 import org.lwjgl.input.Mouse;
+
+import java.awt.*;
+
+import static org.lwjgl.opengl.GL11.*;
 
 public class SimpleSliderUI extends AbstractComponentUI<Slider>
 {
@@ -48,7 +44,7 @@ public class SimpleSliderUI extends AbstractComponentUI<Slider>
 			case INTEGER:
 				content =
 					String.format("%,d",
-						Long.valueOf(Math.round(component.getValue())));
+							Math.round(component.getValue()));
 				break;
 			case PERCENTAGE:
 				int percent =
